Refactor: 자전거 카드 및 제품 정보 섹션에서 태그 색상 처리 개선 - getTagColor 유틸리티 함수 추가 및 코드 정리
