+++ src/utils/getTagColor.ts
@@ -1,0 +1,14 @@
+const getTagColor = (color: string) => {
+  const colorMap: Record<string, string> = {
+    red: "bg-red-500",
+    green: "bg-green-500",
+    blue: "bg-blue-500",
+    yellow: "bg-yellow-500",
+    purple: "bg-purple-500",
+    orange: "bg-orange-500",
+    gray: "bg-gray-500",
+  };
+  return colorMap[color] || "bg-gray-500";
+};
+
+export default getTagColor;
